refactor :User Entity를 auth 패키지에서 user 패키지로 변경
