+++ src/main/kotlin/com/flab/ticketing/auth/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.flab.ticketing.auth.repository
 
-import com.flab.ticketing.auth.entity.User
+import com.flab.ticketing.user.entity.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
